backend/requirements.txt
fastapi
uvicorn[standard]
web3
eth-account
httpx
pyyaml
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-aws>=0.3.0
langchain-google-genai>=1.0.0
langchain-core>=0.3.0
langgraph>=0.2.0
python-dotenv
pydantic>=2.0.0
requests
boto3
